move Push view to git_real
